requirements.txt
fastapi
uvicorn
pydantic
pydantic[email]
python-multipart
dotenv
pymongo
streamlit
bcrypt
pyjwt